third_party/blink/web_tests/http/tests/websocket/receive-blob-expected.txt

WebSocket: Receive Blobs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS receivedMessages.length is 3
Checking message #0.
PASS responseType is "[object Blob]"
PASS actualArray.length is 13
PASS Passed: Message #0.
Checking message #1.
PASS responseType is "[object Blob]"
PASS actualArray.length is 0
PASS Passed: Message #1.
Checking message #2.
PASS responseType is "[object Blob]"
PASS actualArray.length is 256
PASS Passed: Message #2.
PASS successfullyParsed is true

TEST COMPLETE
